refactor: 修改maxGroupMemberCount属性类型[Integer->Long] style: 格式化代码。

+++ src/main/java/com/r/chat/entity/vo/SysSettingVO.java
@@ -12,20 +12,43 @@ import java.io.Serializable;
 public class SysSettingVO implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    // 个人最多群聊数量
+    /**
+     * 个人最多群聊数量
+     */
     private Integer maxGroupCount;
-    // 单个群聊最多群成员
-    private Integer maxGroupMemberCount;
-    // 图片最大大小
+
+    /**
+     * 单个群聊最多群成员
+     */
+    private Long maxGroupMemberCount;
+
+    /**
+     * 图片最大大小
+     */
     private Integer maxImageSize;
-    // 视频最大大小
+
+    /**
+     * 视频最大大小
+     */
     private Integer maxVideoSize;
-    // 文件最大大小
+
+    /**
+     * 文件最大大小
+     */
     private Integer maxFileSize;
-    // 初始机器人id
+
+    /**
+     * 初始机器人id
+     */
     private String robotId;
-    // 初始机器人昵称
+
+    /**
+     * 初始机器人昵称
+     */
     private String robotNickName;
-    // 初始机器人欢迎语
+
+    /**
+     * 初始机器人欢迎语
+     */
     private String robotWelcomeMsg;
 }
